requirements.txt
# ACIS Trading Platform Dependencies

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0

# Scheduling & Automation
APScheduler>=3.10.0
pytz>=2023.3

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP Requests & APIs
requests>=2.31.0
httpx>=0.24.0

# Logging & Monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Data Validation & Types
marshmallow>=3.20.0
marshmallow-sqlalchemy>=0.29.0

# Caching
redis>=4.6.0
cachetools>=5.3.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: Financial Data APIs
yfinance>=0.2.0
alpha-vantage>=2.3.1

# Optional: Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0

# Utilities
click>=8.1.0
tqdm>=4.65.0
python-dateutil>=2.8.0